Big refactoring per comments below:
I want the gui in the preferences to show the amount selected by the user (XX) out of the total amount of memory that I can assign (XX/total).  For 32 bit OS, the total should be the minimum of the (total system memory - 512Mb) and 1536MB.  On 64-bit systems, it should be the total system memory-512Mb.  (512 is arbitrary, but should be enough for the OS).  The default when the user first starts up should be to use 80% of the total, but they should be allowed to increase that to 100% of the total.  The minimum amount they can select should be 512.