Simplifying logic and being extra-conservative on 32 bit:
The default when the user first starts up should be to use 80% between the "low and high-water mark", but they should be allowed to increase that to 100% of the high-water mark. There are a couple of exceptions to the 80% heuristic. The amount of memory returned will never be < 512GB.
Also, if on 32 bit OS be conservative and choose 70%. For example, if we have 32 bit windows with 1536 of memory, the result will be 1229.

The user can "manually" increase to 1536, if they want.