Refactor initReadout for component group compatibility.
This reworks the logic within initReadout so that setReadoutFormat is
always called, even if the IdvWindow passed to initReadout is null (as
seems to be the case when dealing with component groups).

Always calling setReadoutFormat like this is safe because it always
does an internal check for whether or not the readout field is null.