Restore user's selected logo position.
If a user does something like:

1. Open the "View>Properties" window.
2. Select a screen position other than "Lower Left".
3. Click "OK" to close the window.
4. Reopen the "View>Properties" window.

Prior to this commit, the screen position will be reset to "Lower
Left". This commit simply remembers the user's selection by storing it
in a new ViewManager String field ("selectedLogoPosition").

Since the selection is just an instance field, the selected value is
only remembered for the lifetime of the ViewManager and each
ViewManager can have a different selected value (but this seems to
conform pretty well with how things “should” work?).

This is a fix for Inquiry 1219:

http://mcidas.ssec.wisc.edu/inquiry-v/?inquiry=1219