Fix problem with leaving full screen mode.
All this commit does is explicitly run the logic of “resetFullScreen”
on the EDT via invokeLater.

I did test things with GuiUtils.invokeInSwingThread and it seemed to
work, but I opted for invokeLater because:

1. GuiUtils.invokeInSwingThread seems semantically equivalent to always
using SwingUtilities.invokeAndWait (and that shouldn’t be called from
the EDT).

2. Adding the runnable to the end of the EDTs queue—rather than
executing the runnable immediately—seems like the safer choice, as
there could be situations where events/threads reference the
fullScreenWindow JFrame that has already been disposed and nulled out.