Two modifications to showOkCancelDialog.
This adds a parameter named "resizable" to a new showOkCancelDialog
method that controls whether or not the resulting dialog can be resized.

The other change is that the now the resulting dialog's preferred size
is set to the preferred size of the "contents" parameter.

Existing calls in the IDV to "showOkCancelDialog" should behave the same
as before, as the default behavior is to set "resizable" to true.